# docs/swagger.yaml
basePath: /
definitions:
  controllers.AddCubesRequest:
    properties:
      cubes_to_add:
        example: 10
        type: integer
    type: object
...
  controllers.UpdateTokensRequest:
    properties:
      dfc_balance:
        type: number
      m5_balance:
        type: number
      ton_balance:
        type: number
    type: object
  controllers.Withdrawal:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      jetton_name:
        type: string
      wallet:
        type: string
    type: object
...
  entities.User:
    properties:
      created_at:
        type: string
      cubes:
        type: integer
      dfc_balance:
        type: number
      first_name:
        description: Новое поле FirstName
        type: string
      id:
        type: string
      language:
        type: string
      m5_balance:
        type: number
      name:
        type: string
      points:
        type: number
      referral_code:
        type: string
      referral_earnings:
        additionalProperties:
          type: number
        type: object
      referred_by:
        type: string
      tgid:
        type: string
      ton_balance:
        type: number
      updated_at:
        type: string
      wallet:
        type: string
    type: object
...
host: api.m5dice.com
info:
  contact: {}
  description: API for managing users and games in TG-Dice application
  title: TG-Dice API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Retrieve a list of users with optional pagination
      parameters:
      - description: Number of users to retrieve
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List users with pagination
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Для создания Необходим только валлет
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entities.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user from the system by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user by ID
      tags:
      - users
    get:
      description: Retrieve a user by their unique ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - users
  /users/{wallet}/balances:
    get:
      description: Get token balances (ton_balance, m5_balance, dfc_balance) and cubes
        for a user by their wallet
      parameters:
      - description: User Wallet
        in: path
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Balances (ton_balance, m5_balance, dfc_balance, cubes)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user balances
      tags:
      - users
  /users/{wallet}/cubes:
    patch:
      consumes:
      - application/json
      description: Increment the number of cubes for a user by their wallet
      parameters:
      - description: User Wallet
        in: path
        name: wallet
        required: true
        type: string
      - description: Number of cubes to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.AddCubesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add cubes to a user
      tags:
      - users
  /users/{wallet}/points:
    get:
      description: Retrieve the points of a user by their wallet
      parameters:
      - description: User Wallet
        in: path
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User Points
          schema:
            additionalProperties:
              type: number
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user points by wallet
      tags:
      - users
  /users/{wallet}/referral-code:
    get:
      consumes:
      - application/json
      description: Retrieve the referral code for a user by their wallet
      parameters:
      - description: User Wallet
        in: path
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get referral code by user wallet
      tags:
      - users
  /users/{wallet}/referral-earnings:
    get:
      description: Retrieve all referral earnings (ton_balance, m5_balance, dfc_balance)
        for a user by their wallet
      parameters:
      - description: User Wallet
        in: path
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Referral earnings (ton_balance, m5_balance, dfc_balance)
          schema:
            additionalProperties:
              type: number
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get referral earnings by user wallet
      tags:
      - users
  /users/{wallet}/tokens:
    patch:
      consumes:
      - application/json
      description: Update the balances of one or more tokens for a user by their wallet
      parameters:
      - description: User Wallet
        in: path
        name: wallet
        required: true
        type: string
      - description: Token balances to update (e.g., {\
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateTokensRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user token balances
      tags:
      - users
  /users/{wallet}/withdrawals:
    get:
      description: Retrieve a list of withdrawals for a user by their wallet address
      parameters:
      - description: User Wallet
        in: path
        name: wallet
        required: true
        type: string
      - description: Number of withdrawals to retrieve
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of withdrawals
          schema:
            items:
              $ref: '#/definitions/controllers.Withdrawal'
            type: array
        "500":
          description: Error fetching withdrawals
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get withdrawals by user wallet
      tags:
      - users
  /users/name/{name}:
    get:
      description: Retrieve a user by their unique name
      parameters:
      - description: User Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by name
      tags:
      - users
  /users/points:
    get:
      description: Retrieve a list of users sorted by their points in descending order
      parameters:
      - description: Number of users to retrieve
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get users sorted by points
      tags:
      - users
  /users/tgid/{tgid}:
    patch:
      consumes:
      - application/json
      description: Update only specific fields of a user identified by TgID
      parameters:
      - description: Telegram ID
        in: path
        name: tgid
        required: true
        type: string
      - description: Fields to update
        in: body
        name: updateData
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated fields
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Partially update a user by TgID
      tags:
      - users
...
schemes:
- https
swagger: "2.0"